Extract #each to object (sometimes self)
