+++ app/workers/concerns/dump_commons.rb
@@ -4,6 +4,7 @@ module DumpCommons
     @csv_dump ||= CsvDump.new
   end
 
+
   def upload_dump
     gzip_file_path = csv_dump.gzip!
     write_to_s3(gzip_file_path)
