+++ app/workers/emails_projects_export_worker.rb
@@ -7,12 +7,12 @@ class EmailsProjectsExportWorker
 
   sidekiq_options queue: :data_low
 
-  attr_reader :project
-
   def perform(project_id)
     @project = Project.find(project_id)
+    @scope = CsvDumps::ProjectEmailList.new(project_id)
+
     begin
-      emails_to_csv_file
+      perform_dump
       upload_dump
     ensure
       cleanup_dump
@@ -24,24 +24,7 @@ class EmailsProjectsExportWorker
     @formatter ||= Formatter::Csv::UserEmail.new
   end
 
-  def each
-    read_from_database do
-      user_emails.find_each do |user|
-        yield user
-      end
-    end
-  end
-
   private
-
-  def user_emails
-    User
-      .joins(:project_preferences)
-      .where(user_project_preferences: { project_id: project.id, email_communication: true })
-      .active
-      .where(valid_email: true)
-      .select(:id, :email)
-  end
 
   def export_type
     project.slug
