+++ app/workers/emails_users_export_worker.rb
@@ -11,6 +11,7 @@ class EmailsUsersExportWorker
 
   def perform(export_type=:global)
     @export_type = export_type
+    @scope = CsvDumps::GenericEmailList.new(export_type)
     begin
       perform_dump
       upload_dump
@@ -21,32 +22,5 @@ class EmailsUsersExportWorker
 
   def formatter
     @formatter ||= Formatter::Csv::UserEmail.new
-  end
-
-  def each
-    read_from_database do
-      user_emails.find_each do |user|
-        yield user
-      end
-    end
-  end
-
-  private
-
-  def user_emails
-    export_field = case export_type
-    when :global
-      :global_email_communication
-    when :beta
-      :beta_email_communication
-    end
-    emailable_users(export_field).select(:id, :email, export_field)
-  end
-
-  def emailable_users(export_field)
-    @emailable_users ||= User
-      .active
-      .where(valid_email: true)
-      .where(export_field => true)
   end
 end
